Revert "feat: 가장 최근 문제를 푼 달을 구하는 함수를 유틸로 분리"
This reverts commit 915ccf2039272e5cec5f6a750bc4bdacab66b41a.
